# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillBridge AI",
    "description": "A free education and professional development service for military members and veterans. Modular monolith with an isolated AI Integration Module, deterministic escalation rules and asynchronous feedback analysis. Every error uses one envelope: `{\"error\": {\"code\", \"message\", \"requestId\"}}`. The written contract is docs/API.md.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/profile/record/items": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Add Item",
        "operationId": "add_item_api_v1_profile_record_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRecordItemView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profile/record/items/bulk": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Add Items",
        "description": "Save the items the member confirmed after an upload. Repeats are counted, not errors.",
        "operationId": "add_items_api_v1_profile_record_items_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordItemsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordItemsAddedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profile/record/items/{item_id}": {
      "delete": {
        "tags": [
          "profile"
        ],
        "summary": "Remove Item",
        "operationId": "remove_item_api_v1_profile_record_items__item_id__delete",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profile/record/import": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Import Record",
        "description": "Read an uploaded .txt, .csv or .pdf and suggest items. Saves nothing.",
        "operationId": "import_record_api_v1_profile_record_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberRecordItemView": {
        "properties": {
          "itemId": {
            "type": "string",
            "format": "uuid",
            "title": "Itemid"
          },
          "kind": {
            "$ref": "#/components/schemas/RecordItemKind"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source": {
            "$ref": "#/components/schemas/RecordItemSource"
          },
          "addedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Addedat"
          }
        },
        "type": "object",
        "required": [
          "itemId",
          "kind",
          "name",
          "source",
          "addedAt"
        ],
        "title": "MemberRecordItemView"
      },
...
      "RecordCandidate": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/RecordItemKind"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "name"
        ],
        "title": "RecordCandidate"
      },
      "RecordImportRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "contentBase64": {
            "type": "string",
            "title": "Contentbase64"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "contentBase64"
        ],
        "title": "RecordImportRequest"
      },
      "RecordImportResponse": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/RecordCandidate"
            },
            "type": "array",
            "title": "Candidates"
          },
          "skippedLines": {
            "type": "integer",
            "title": "Skippedlines"
          }
        },
        "type": "object",
        "required": [
          "candidates",
          "skippedLines"
        ],
        "title": "RecordImportResponse"
      },
      "RecordItemKind": {
        "type": "string",
        "enum": [
          "TRAINING",
          "CREDENTIAL"
        ],
        "title": "RecordItemKind"
      },
      "RecordItemRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "name"
        ],
        "title": "RecordItemRequest"
      },
      "RecordItemSource": {
        "type": "string",
        "enum": [
          "MANUAL",
          "UPLOAD"
        ],
        "title": "RecordItemSource"
      },
      "RecordItemsAddedResponse": {
        "properties": {
          "added": {
            "items": {
              "$ref": "#/components/schemas/MemberRecordItemView"
            },
            "type": "array",
            "title": "Added"
          },
          "alreadyOnRecord": {
            "type": "integer",
            "title": "Alreadyonrecord"
          }
        },
        "type": "object",
        "required": [
          "added",
          "alreadyOnRecord"
        ],
        "title": "RecordItemsAddedResponse"
      },
      "RecordItemsRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecordItemRequest"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "RecordItemsRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}